requirements.txt
asgiref==3.11.1
Django==5.2.11
sqlparse==0.5.5
typing_extensions==4.15.0
dj-database-url
python-dotenv
djangorestframework
djangorestframework-simplejwt